Version 0.3.1. - Changed C++ standard from C++11 to C++14 (it's time to start using std::make_unique). - Fixed the crash occurring when showing an animation from CC UOP. - Added const_iterator to ScriptObjTree. - Tidied the ScriptSearch/SubDlg_SearchObj code. - Changed some unscoped enums to enum class. - Changed UOIdx::unknown to extra.
UOPPackage library version 3.2.
- Added iterators to UOPPackage and UOPBlock.
- Changed some data types from signed to unsigned (block&file number, addresses).
- Added UOPFile::freeData and hasData methods.
- Removed unused UOPHeader.cpp/.h.

+++ src/keystrokesender/keystrokesender.cpp
@@ -5,7 +5,7 @@ namespace ks
 {
 
 char const * getErrorStringStatic(KSError err) {
-    return KSErrorString[err];
+    return KSErrorString[(int)err];
 }
 
 void KeystrokeSender::setSetFocusToWindow(bool value) {
@@ -17,7 +17,7 @@ KSError KeystrokeSender::getError() const {
 }
 
 std::string KeystrokeSender::getErrorString() const {
-    return KSErrorString[m_error];
+    return KSErrorString[(int)m_error];
 }
 UOClientType KeystrokeSender::getClientType() const {
     return m_clientType;
